Refactor disruptor package
