+++ pkg/disruptors/disruptor.go
@@ -1,0 +1,7 @@
+package disruptors
+
+// Disruptor defines the generic interface implemented by all disruptors
+type Disruptor interface {
+	// Targets returns the list of targets for the disruptor
+	Targets() ([]string, error)
+}
