+++ pkg/disruptors/protocol.go
@@ -2,6 +2,32 @@ package disruptors
 
 import "time"
 
+// ProtocolFaultInjector defines the methods for injecting protocol faults
+type ProtocolFaultInjector interface {
+	// InjectHTTPFault injects faults in the HTTP requests sent to the disruptor's targets
+	// for the specified duration
+	InjectHTTPFaults(fault HTTPFault, duration time.Duration, options HTTPDisruptionOptions) error
+	// InjectGrpcFault injects faults in the grpc requests sent to the disruptor's targets
+	// for the specified duration
+	InjectGrpcFaults(fault GrpcFault, duration time.Duration, options GrpcDisruptionOptions) error
+}
+
+// HTTPDisruptionOptions defines options for the injection of HTTP faults in a target pod
+type HTTPDisruptionOptions struct {
+	// Port used by the agent for listening
+	ProxyPort uint `js:"proxyPort"`
+	// Network interface the agent will be listening traffic from
+	Iface string
+}
+
+// GrpcDisruptionOptions defines options for the injection of grpc faults in a target pod
+type GrpcDisruptionOptions struct {
+	// Port used by the agent for listening
+	ProxyPort uint `js:"proxyPort"`
+	// Network interface the agent will be listening traffic from
+	Iface string
+}
+
 // HTTPFault specifies a fault to be injected in http requests
 type HTTPFault struct {
 	// port the disruptions will be applied to
